Move models inside api folder
