+++ src/app/api/tests/test_categories.py
@@ -4,7 +4,7 @@ from dotenv import load_dotenv
 from flask import json
 
 from app import create_app, db
-from app.models.category import Category
+from app.api.models.category import Category
 
 
 @unittest.skip("Waiting to be implemented.")
