+++ src/app/api/tests/test_questions.py
@@ -4,9 +4,9 @@ from dotenv import load_dotenv
 from flask import json
 
 from app import create_app, db
-from app.models.answer import Answer
-from app.models.category import Category
-from app.models.question import Question
+from app.api.models.answer import Answer
+from app.api.models.category import Category
+from app.api.models.question import Question
 
 
 @unittest.skip("Waiting to be implemented.")
